# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is a sample recipes API. You can find out more about the API at https://github.com/faishalshidqi/gin-introductory-proj",
    "title": "Recipes API",
    "contact": {
      "name": "Faishal Shidqi",
      "email": "faishalshidqi.work@proton.me"
    },
    "version": "1.0.0"
  },
  "host": "localhost:9000",
  "paths": {
    "/recipes": {
      "get": {
        "description": "Returns list of recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "retrieveRecipes"
      }
    },
    "/recipes/": {
      "post": {
        "description": "Create a new recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "addRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "name of the recipe",
            "name": "name",
            "in": "body",
            "required": true
          },
          {
            "description": "tags of the recipe",
            "name": "tags",
            "in": "body",
            "required": true
          },
          {
            "description": "ingredients of the recipe",
            "name": "ingredients",
            "in": "body",
            "required": true
          },
          {
            "description": "instructions to make the recipe",
            "name": "instructions",
            "in": "body",
            "required": true
          }
        ]
      }
    },
    "/recipes/:id": {
      "get": {
        "description": "Look for recipe with given id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "getOneRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "path of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      }
    },
    "/recipes/search": {
      "get": {
        "description": "Look for recipe(s) with given tag",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "searchRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "tag of the recipe",
            "name": "tag",
            "in": "query",
            "required": true
          }
        ]
      }
    },
    "/recipes/{id}": {
      "put": {
        "description": "Update an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "updateRecipe",
        "parameters": [
          {
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "name of the recipe",
            "name": "name",
            "in": "body",
            "required": true
          },
          {
            "description": "tags of the recipe",
            "name": "tags",
            "in": "body",
            "required": true
          },
          {
            "description": "ingredients of the recipe",
            "name": "ingredients",
            "in": "body",
            "required": true
          },
          {
            "description": "instructions to make the recipe",
            "name": "instructions",
            "in": "body",
            "required": true
          }
        ]
      },
      "delete": {
        "description": "Delete a recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "removeRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      }
    }
  }
}